# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server Petstore server.",
        "title": "Swagger Example API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "petstore.swagger.io",
    "basePath": "/v2",
    "paths": {
        "/v1/toys/": {
            "post": {
                "description": "create toy",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Create toy",
                "parameters": [
                    {
                        "description": "新建玩具body入参",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.ToyReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/v1/toys/{toyID}": {
            "get": {
                "description": "get toy by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Get toy",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "path参数：玩具ID",
                        "name": "toyID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "query参数：页码",
                        "name": "pageNum",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "每页数量",
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "header参数：request id",
                        "name": "requestID",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "types.ToyReq": {
            "type": "object",
            "properties": {
                "name": {
                    "description": "the name for this toy\nrequired: true\nmin length: 3",
                    "type": "string"
                },
                "price": {
                    "description": "the price for this toy\nrequired: true\nmin: 1",
                    "type": "integer"
                }
            }
        }
    }
}